Seperated student class.
